# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Gloo",
        "title": "Gloo API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@gloo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "gloo-api-production.up.railway.app",
    "basePath": "/api",
    "paths": {
...
        "/instructions/{recipeId}/{instructionId}": {
            "get": {
                "description": "Get a specific instruction by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Instructions"
                ],
                "summary": "Get instruction by ID",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Recipe ID",
                        "name": "recipeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Instruction ID",
                        "name": "instructionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved instruction",
                        "schema": {
                            "$ref": "#/definitions/models.GetInstructionByIdResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Instruction not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a specific instruction by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Instructions"
                ],
                "summary": "Update instruction",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Recipe ID",
                        "name": "recipeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Instruction ID",
                        "name": "instructionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Instruction",
                        "name": "instruction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateInstructionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated instruction",
                        "schema": {
                            "$ref": "#/definitions/models.UpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Instruction not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new instruction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Instructions"
                ],
                "summary": "Create instruction",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Recipe ID",
                        "name": "recipeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Instruction ID",
                        "name": "instructionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Instruction",
                        "name": "instruction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateInstructionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created instruction",
                        "schema": {
                            "$ref": "#/definitions/models.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Instruction not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a specific instruction by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Instructions"
                ],
                "summary": "Delete instruction",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Recipe ID",
                        "name": "recipeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Instruction ID",
                        "name": "instructionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted instruction",
                        "schema": {
                            "$ref": "#/definitions/models.DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Instruction not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/notifications/{userId}": {
            "get": {
                "description": "Get all notifications for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Get all notifications",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetAllNotificationsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new notification",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "Create a notification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Notification",
                        "name": "notification",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateNotificationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/notifications/{userId}/{notificationId}": {
            "get": {
                "description": "Get a notification by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Get notification by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Notification ID",
                        "name": "notificationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetNotificationsByIdResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateInstructionRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Chop the chicken into small cubes or finely mince it (you can also use a food processor). This makes it easier to form patties later."
                },
                "image_url": {
                    "type": "string",
                    "example": "https://example.com/image.png"
                },
                "step": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
        "models.CreateNotificationRequest": {
            "type": "object",
            "properties": {
                "actor_id": {
                    "type": "integer",
                    "example": 1
                },
                "created_by": {
                    "type": "string",
                    "example": "@facu.potti"
                },
                "description": {
                    "type": "string",
                    "example": "A feast for the senses"
                },
                "is_new": {
                    "type": "boolean",
                    "example": true
                },
                "title": {
                    "type": "string",
                    "example": "Chicken Wings"
                },
                "type": {
                    "type": "string",
                    "example": "follow"
                }
            }
        },
...
        "models.CreateResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Created successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "status_code": {
                    "type": "integer",
                    "example": 201
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.DeleteResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Deleted successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.ErrorBadRequestResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Bad Request"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                },
                "status_code": {
                    "type": "integer",
                    "example": 400
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.ErrorInternalServerErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Internal Server Error"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                },
                "status_code": {
                    "type": "integer",
                    "example": 500
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.ErrorNotFoundResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Not Found"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                },
                "status_code": {
                    "type": "integer",
                    "example": 404
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "models.GetAllNotificationsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "actor_id": {
                                "type": "integer",
                                "example": 1
                            },
                            "created_by": {
                                "type": "string",
                                "example": "@facu.potti"
                            },
                            "description": {
                                "type": "string",
                                "example": "A feast for the senses"
                            },
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "is_new": {
                                "type": "boolean",
                                "example": true
                            },
                            "title": {
                                "type": "string",
                                "example": "Chicken Wings"
                            },
                            "type": {
                                "type": "string",
                                "example": "follow"
                            }
                        }
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Successfully retrieved notifications"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.GetInstructionByIdResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "description": {
                            "type": "string",
                            "example": "Chop the chicken into small cubes or finely mince it (you can also use a food processor). This makes it easier to form patties later."
                        },
                        "id": {
                            "type": "integer",
                            "example": 1
                        },
                        "image_url": {
                            "type": "string",
                            "example": "https://example.com/image.png"
                        },
                        "ingredients": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "onion"
                                    },
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Chicken"
                                    },
                                    "quantity": {
                                        "type": "string",
                                        "example": "100"
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "kg"
                                    }
                                }
                            }
                        },
                        "step": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Successfully retrieved instruction"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.GetNotificationsByIdResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "actor_id": {
                            "type": "integer",
                            "example": 1
                        },
                        "created_by": {
                            "type": "string",
                            "example": "@facu.potti"
                        },
                        "description": {
                            "type": "string",
                            "example": "A feast for the senses"
                        },
                        "id": {
                            "type": "integer",
                            "example": 1
                        },
                        "is_new": {
                            "type": "boolean",
                            "example": true
                        },
                        "title": {
                            "type": "string",
                            "example": "Chicken Wings"
                        },
                        "type": {
                            "type": "string",
                            "example": "follow"
                        }
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Successfully retrieved notification"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.UpdateInstructionRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Chop the chicken into small cubes or finely mince it (you can also use a food processor). This makes it easier to form patties later."
                },
                "image_url": {
                    "type": "string",
                    "example": "https://example.com/image.png"
                },
                "step": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
...
        "models.UpdateResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Updated successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    }
}